update Mario Kart Advance
added some missing data to most files.
converted mobile GP rule values to hex (that is, allowed a mobile GP to start the player with more than 45 coins).
also added a query mark to index.html's img src, so the PHP development server will accept the request for it. in this commit because i'm not great at using Git(Hub) properly.